Add copy on click for type values
